Added point system where if the ball reaches horizontal ends of the screen, a goal gets added and the ball resets to the middle.
Modified the ballMovement function to return 2000 if the ball surpasses the right boundary and -100 if it surpasses the left boundary.

In the main function, if the 3rd index of ballCoordinates is 2000 or -100 it will know there was a goal and adds the score. It also resets the ball to the middle.