Fixed the issue on the right bar where the bounce when the ball hit the middle of the bar cause the ball to go to extreme angles (e.g. 80 degrees). This was because I was inverting the x direction of the bounce by summing pi/2 radians (90 degrees) and instead to invert I needed to subtract the to pi the bounce angle.
Corrected the issue where the bounce on the top corner of the bars caused the ball to go to an angle above the maxBounceAngle set (above 75 degrees). This was caused by the math not taking into account that the ball origin is in the top left corner of the ball, so when it hit the top corner is was length/2 distance from the center of bar +12 which is the height of the ball. This was fixed by setting the midpoint of the bar 56 pixels from the origin (100px of the bar + 12px of the ball)/2.